+++ lib/statistics.d
@@ -8,8 +8,8 @@ import std.random: dice;
 // Mean value of a time series
 double compute_mean(double[] array_IN) pure {
 	double sum = 0;
-	ulong n = array_IN.length;
-	for(ulong i = 0; i<n; i++) { sum += array_IN[i]; }
+	int n = array_IN.length;
+	for(int i = 0; i<n; i++) { sum += array_IN[i]; }
 	return sum/n;
 }
 
@@ -17,8 +17,8 @@ double compute_mean(double[] array_IN) pure {
 double compute_std(double[] array_IN) pure {
 	double sum = 0;
 	double mean = compute_mean(array_IN);
-	ulong n = array_IN.length;
-	for(ulong i = 0; i<n; i++) { sum += pow((array_IN[i]-mean),2); }
+	int n = array_IN.length;
+	for(int i = 0; i<n; i++) { sum += pow((array_IN[i]-mean),2); }
 	return sqrt(sum/n);	
 }
 
@@ -31,12 +31,12 @@ double compute_variance(double[] array_IN) pure {
 // Correlation between two time series.
 double compute_correlation(double[] array_x_IN, double[] array_y_IN) pure {
 	enforce(array_x_IN.length == array_y_IN.length, "compute_correlation: not same length"); 
-	ulong n = array_x_IN.length;
+	int n = array_x_IN.length;
 	double x_mean = compute_mean(array_x_IN);
 	double y_mean = compute_mean(array_y_IN);
 	double numerator = 0.0;
 	double denominator = 0.0;
-	for (ulong i = 0; i<n; i++) {
+	for (int i = 0; i<n; i++) {
 		numerator += (array_x_IN[i]-x_mean)*(array_y_IN[i]-y_mean);
 		denominator += sqrt(pow(array_x_IN[i]-x_mean, 2)*pow(array_y_IN[i]-y_mean, 2));
 	}
@@ -47,13 +47,13 @@ double compute_correlation(double[] array_x_IN, double[] array_y_IN) pure {
 // y = b1 + b2*x
 double[2] compute_ordinary_least_squares(double[] array_x_IN, double[] array_y_IN) pure {
 	enforce(array_x_IN.length == array_y_IN.length, "compute_ordinary_least_squares: not same length"); 
-	ulong n = array_x_IN.length;
+	int n = array_x_IN.length;
 	double x_mean = compute_mean(array_x_IN);
 	double y_mean = compute_mean(array_y_IN);
 	double numerator = 0.0;
 	double denominator = 0.0;
 	double b1, b2;
-	for(ulong i = 0; i<n; i++) {
+	for(int i = 0; i<n; i++) {
 		numerator += (array_x_IN[i]-x_mean)*(array_y_IN[i]-y_mean);
 		denominator += pow(array_x_IN[i]-x_mean, 2);
 	}
